+++ groups/bdl/bdlma/bdlma_infrequentdeleteblocklist.t.cpp
@@ -746,7 +746,7 @@ int main(int argc, char *argv[])
 
             const void *EXP_P = (char *)oa.lastAllocatedAddress() + HDRSZ;
 
-            int numBytes = oa.lastAllocatedNumBytes();
+            int numBytes = static_cast<int>(oa.lastAllocatedNumBytes());
 
             int offset = U::calculateAlignmentOffset(p, U::BSLS_MAX_ALIGNMENT);
